+++ CaskHub/Services/Catalog/CategoryService.swift
