+++ CaskHub/Services/Downloads/DownloadMetadataProvider.swift
