+++ CaskHub/Services/Homebrew/Coordination/CaskLocalStateResolver.swift
