+++ CaskHub/Services/Homebrew/Coordination/CaskOperationFailureFactory.swift
