+++ CaskHub/Services/Homebrew/Coordination/CaskOperationStore.swift
