+++ CaskHub/Services/Homebrew/Coordination/InstallationCatalogBuilder.swift
