+++ CaskHub/Services/Homebrew/Domain/CaskOperationProgress.swift
