+++ CaskHub/Services/Homebrew/Domain/CaskOperationState.swift
