+++ CaskHub/Services/Homebrew/Domain/InstallReceipt.swift
