+++ CaskHub/Services/Homebrew/Domain/InstallationSnapshot.swift
