+++ CaskHub/Services/Homebrew/Domain/LocalHomebrewTypes.swift
