+++ CaskHub/Services/Homebrew/Facade/LocalHomebrewService+Adoption.swift
