+++ CaskHub/Services/Homebrew/Facade/LocalHomebrewService+Intents.swift
