+++ CaskHub/Services/Homebrew/Facade/LocalHomebrewService+Snapshot.swift
