+++ CaskHub/Services/Homebrew/Facade/LocalHomebrewService+State.swift
