+++ CaskHub/Services/Homebrew/Infrastructure/AppManagementPermission.swift
