+++ CaskHub/Services/Homebrew/Infrastructure/ApplicationDiscovery.swift
