+++ CaskHub/Services/Homebrew/Infrastructure/ApplicationLauncher.swift
