+++ CaskHub/Services/Homebrew/Infrastructure/AskpassScriptManager.swift
