+++ CaskHub/Services/Homebrew/Infrastructure/BrewOutputCollector.swift
