+++ CaskHub/Services/Homebrew/Infrastructure/HomebrewCommandExecutor.swift
