+++ CaskHub/Services/Homebrew/Infrastructure/HomebrewInstallationScanner.swift
