+++ CaskHub/Services/Homebrew/Infrastructure/HomebrewOutputDiagnostics.swift
