+++ CaskHub/Services/Homebrew/Infrastructure/InstallationIndexBuilder.swift
