+++ CaskHub/Services/Homebrew/Infrastructure/PackageReceiptResolver.swift
