+++ CaskHub/ViewModels/Catalog/CaskCatalogViewModel+Projection.swift
