+++ CaskHub/ViewModels/Catalog/Projection/CatalogProjectionCache.swift
