+++ CaskHub/ViewModels/Catalog/Projection/CatalogProjector.swift
